fix: move notification constant to constant folder
